auth_service/pyproject.toml
# pyproject.toml для Auth Service
# Менеджер пакетов: uv (https://docs.astral.sh/uv/)
# Установка зависимостей: uv sync

[project]
name = "auth-service"
version = "0.1.0"
description = "Auth service: registration + JWT issuing"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "aiosqlite>=0.20.0",
  "pydantic[email]>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "ruff>=0.14.0",
  "greenlet>=3.3.1",
  "bcrypt==4.3.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "respx>=0.21.0",
]

[tool.uv]
dev-dependencies = []
